move relationship class
