+++ urbanstats/website_data/build.py
@@ -4,7 +4,7 @@ import shutil
 
 from urbanstats.website_data.output_geometry import produce_all_geometry_json
 from urbanstats.website_data.create_article_gzips import create_article_gzips, extra_stats
-from relationship import map_relationships_by_type, ordering_idx as type_ordering_idx, type_to_type_category
+from urbanstats.geometry.relationship import map_relationships_by_type, ordering_idx as type_ordering_idx, type_to_type_category
 from urbanstats.geometry.shapefiles.shapefiles_list import american_to_international
 from urbanstats.consolidated_data.produce_consolidated_data import full_consolidated_data, output_names
 from urbanstats.mapper.ramp import output_ramps
